Moved spotifyLinkUtil into spotify folder
